Update call git clean on debian/rules
When making build sairedis directory could be polluted by previous build
which was configured without '--enable-syncd=no', and that created
makefiles for previous build for example binary-syncd target, next when
dh clean is executed for target binary-sairedis, it's executing existing
configure with previous parameters by executing "./bin/bash
./config.status .recheck", this don't contain --enable-syncd=no, and
this requires libsai to be present for configuration, this will cause
configure to fail, but dh clean will ignore that error and continue next
commands. Because of this, make distclean was not executed, and
directory contains previous build artifacts like _pysairedis.la in pyext
directory, which was from stretch build, and points to different python
version 3.5 vs 3.7 on buster, and since _pysairedis.la exists, it's
content are used to pass that python library to link stage of python
build.

To workaround this, we can export shell variable to disable libsai
requirement, and distclean will succeed, but there still be left
artifacts in syncd directory and others conditioned in Makefile.am

Signed-off-by: kcudnik <kcudnik@gmail.com>