Add to save sai objects which are created during onPostPortCreate
The ports which is without QUEUE related configurations still created related
SCHEDULER_GROUPs in the onPostPortCreate() during the syncd startup in some
platform - e.g. barefoot montara. But these created objects did NOT be recorded
in the `COLDVIDS` and `m_warmBootDiscoveredVids` set. While swss is perform
warmstart, the `compareViews()` would be called before apply the view, and it
would detect the differences between the current view and temp view. Then syncd
would remove these schedulers groups in the finalize.

This patch adds to record the objects which are created during onPostPortCreate
into the `m_warmBootDiscoveredVids` set.