Fix RPC package build failure due to shell syntax issue
PR #1194 removed the .ONESHELL flag (causing all commands for a target
to be executed in a single shell environment). Because of that,
multi-line shell code blocks aren't parsed correctly.

Since the `syncd_init_common.sh` file should always exist anyways (it's
unconditionally sourced by an init script and a start script), remove
the if-check around it, and always run `sed` when building rpc packages.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>