[syncd] Add workaround for SET operation on SAI_HOSTIF_ATTR_QUEUE
On broadcom platform it may happen on warm boot that new firmware will
support query switch capability SET operation for this attribute, but
firmware is not actually implementing that operation, OA in this case
will add new attribute SAI_HOSTIF_ATTR_QUEUE to create HOSTIF object,
and comparison logic will generate SET operation to update existing
object. If set is not supported, then this will cause syncd to
crash because operation will be not successfull.

We are adding this workaround for time being until this will be
supported, then this needs to be removed.