Manually start and stop rsyslogd
For Bookworm, the init.d service file for rsyslogd is removed, and the
only script to start and stop rsyslogd is the systemd service file.
However, that can only be used if systemd is being used as the init
manager, which isn't the case for our docker containers.

As an alternative, manually start and stop rsyslogd ourselves.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>