Fix SSCI parameter when creating MACsec tunnels on Bookworm
While Bookworm's iproute2 has built-in support for the MACsec XPN
protocols, the SSCI argument is passed in in a different way compared to
the patch that we had. Instead of taking in a number, a hex string is
taken in instead.

Update the code in sairedis to pass in a hex string for SSCI instead of
a number.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>