Use RedisChannel instead of ZeroMQChannel for ClientSai
Ideally, context config would be available here so that that information
about what channel to use can be passed in here. However, that is not
available here. Therefore, for our experiemental purposes, just change
ClientSai to use RedisChannel, so that it uses Redis and can be used
with current syncd runtime.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>