add support for "sai_dbg_generate_dump" API
This update introduces support for the sai_dbg_gen_dump API to generate
a debug dump file by the SAI
The general process is as follows:
The filename is written to configDB.
This action triggers the dbgGenDumpOrch to write the necessary data to ASIC DB.
The syncd process reads from ASIC DB and invokes the SAI.
SAI processes the request and returns a response.
syncd writes the response back to redis.
dbgGenDumpOrch retrieves the response and informs the caller by writing to configDB.
The caller then removes the data from configDB.