Add Flex counters bulk Dash meter stats fetch support in syncd
This PR adds syncd FlexCounter support to periodically fetch
SAI bulk stats for Dash meter buckets and write the stats to
COUNTERS_DB.
Multiple Meter buckets, each with inbound and outbound byte counters,
are internally allocated for each ENI object in the Dash pipeline.
To support this a new DashMeterCounterContext class is derived from
BaseCounterContext that tracks and invokes SAI bulkGetStats for all the
meter bucket objects for each ENI object added to the context.

Reference -
https://github.com/sonic-net/DASH/blob/main/documentation/metering/metering.md

Signed-off-by: Mukesh MV <mukesh@pensando.io>