Fix determining first run on gbsyncd warm-reboot
firstRun being true disallows the start of warm-reboot.
It was set to true when no switch attributes is found under the key of
HIDDEN* in ASIC_DB or GB_ASIC_DB. However, all such attributes are unrelated
to the credo gbsyncd, so hasNoHiddenKeysDefined is always true for gbsyncd.
The fix here is to determine the first run also based
on no switch objects has been created in ASIC_DB or GB_ASIC_DB.