Enhance counter polling implementation to store counter support per intf
Currently, a single, global list of counters is used to denote the supported counters for all interfaces. This is done with the assumption that all interfaces support the same counters.

However, on TH5, this is not the case. The S2 ports/MGMT ports dont support all counters that the usual ports do. This causes no counters for the MGMT ports to be stored to COUNTERS_DB.

Change the counter implementation to allow multiple groups of counter support. Each counter group is a set of supported counters to poll, and each interface will be assigned to a counter group that is representative of its capabilities. Multiple interfaces can be assigned the same counter group if they share the same support.

More detailed implemntation details can be found in the proposal:
https://docs.google.com/document/d/1m9hvmPxFFPCRz5n2uDAvS8KJe3K8Eh0kMoxd8bJlk0k