[meta] do not fail bulk operations if MODE_IGNORE_ERROR (#1647)
bulk operations do not continue when SAI_BULK_OP_ERROR_MODE_IGNORE_ERROR is set and an error occurs. Create new check that considers the mode when validating bulk operations.

fixes a crash in https://github.com/sonic-net/sonic-buildimage/issues/21243 because route entries that are not removed in syncd by bulker are assumed to be removed in orchagent. Then when the nexthops are attempting to be removed, orchagent crashes due to OBJECT_IN_USE.